--- .babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["@babel/preset-react"],
-    "plugins": ["@babel/plugin-proposal-class-properties"]
-}
