+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -12,5 +12,5 @@ A clear and concise description of what you want to happen.
 **Describe alternatives you've considered**
 A clear and concise description of any alternative solutions or features you've considered.
 
-**Additional information's**
-Use a label and assign this issue to someone to get your issue done quickly -->
+**Additional context**
+Add any other context or screenshots about the feature request here.
