+++ docs/index.html
@@ -1,15 +1,39 @@
 <!DOCTYPE html>
 <html lang="fr">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-    <link rel="stylesheet" href="main.css">
-    <title>Web-timeline</title>
-</head>
-<body>
-    <main id="app"></main>
-    <script src="bundle.js"></script>
-</body>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width,initial-scale=1">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <link rel="stylesheet" href="main.css">
+        <title>Evolution of technologies</title>
+    </head>
+    <body>
+        <header>
+            <h1 class="is__main__title">The<br> <span class="is__span__title">Evolution</span> <br>of<br> <span class="is__span__title">technologies</span></h1>
+        </header>
+        <nav class="filter" id="js-stickyMenu">
+            <div class="filter--date">
+                <h3 class="is__subTitle">Date</h3>
+            </div>
+            <div class="filter--cat">
+                <select>
+                    <option value="">Category</option>
+                    <option value="Network">Network</option>
+                    <option value="Launch">Launch</option>
+                    <option value="Science">Science</option>
+                    <option value="Security">Security</option>
+                    <option value="Programming">Programming</option>
+                    <option value="Ai">AI</option>
+                    <option value="Social Media">Social Media</option>
+                    <option value="design">Design</option>
+                </select>
+            </div>
+            <div class="filter--title">
+                <h3 class="is__subTitle">Title</h3>
+            </div>
+            <input type="text" name="" class="filter--search" placeholder="Recherche" id="">
+        </nav>
+        <main id="app"></main>
+    <script src="main.js"></script>
+    </body>
 </html>
