+++ docs/main.css
@@ -1,0 +1,1 @@
+.content,.filter,.is__content__important{display:grid;grid-template-columns:repeat(8,1fr);grid-gap:1em;justify-items:start;align-items:center;padding:1em}.is__btn,.is__btn__important{padding:1em;color:#fff;border-radius:4px;border:none;cursor:pointer;-webkit-transition-duration:.4s;transition-duration:.4s}.content{color:#4083a2}.content--date{grid-column:2/2}.content--cat{grid-column:3/3}.content--more{grid-column:8/8}.content--description,.content--title{grid-column:4/7}.content:nth-child(odd){background-color:#f2f7fa}.content,.filter,.is__content__important{display:grid;grid-template-columns:repeat(8,1fr);grid-gap:1em;justify-items:start;align-items:center;padding:1em}.is__btn,.is__btn__important{padding:1em;color:#fff;border-radius:4px;border:none;cursor:pointer;-webkit-transition-duration:.4s;transition-duration:.4s}.sticky{position:fixed;top:0;width:100%}.sticky+#app{padding-top:65px}.filter{background-color:#4083a2;color:#fff;padding:0 1em;box-shadow:0 3px 7px rgba(0,0,0,.3)}.filter--date{grid-column:2/2}.filter--cat{grid-column:3/3}.filter--title{grid-column:4/7}.filter--search{grid-column:8/8}body{max-width:100%;margin:0;padding:0}.content,.filter,.is__content__important{display:grid;grid-template-columns:repeat(8,1fr);grid-gap:1em;justify-items:start;align-items:center;padding:1em}.is__btn,.is__btn__important{padding:1em;color:#fff;border-radius:4px;border:none;cursor:pointer;-webkit-transition-duration:.4s;transition-duration:.4s}.is__none{display:none}.is__btn{background:#4083a2}.is__btn:hover{background-color:#fff;color:#4083a2;box-shadow:0 1px 4px 0 rgba(0,0,0,.1)}.is__btn__important{background-color:#ff7b0a}.is__btn__important:hover{background:#fff;box-shadow:0 1px 4px 0 rgba(0,0,0,.1);color:#ff7b0a}.is__main__title{text-align:center;font-family:SignPainter;line-height:.7em;font-size:3em;font-weight:700}.is__subTitle{font-size:1.1em}.is__text__content{font-size:1em;font-family:Roboto}.is__span__title{font-family:Hominis;font-size:30px;font-weight:100}.is__content__important{background-color:#ffefde;color:#ff7b0a}header{display:flex;justify-content:center;align-items:center;height:10em;padding:1em 0;background-color:#4d9dc2;color:#fff}
