+++ package.json
@@ -1,39 +1,37 @@
 {
-  "name": "",
+  "name": "crispy-octo-adventure",
   "version": "1.0.0",
-  "description": "Boilerplate for your next web application 🐙 🐙",
-  "main": "",
+  "description": "Octopus is a genus of molluscs of the order Octopods",
+  "main": "index.js",
+  "directories": {
+    "doc": "docs"
+  },
   "scripts": {
-    "test": "",
-    "build": "webpack --mode='production'",
-    "start": "webpack-dashboard webpack-dev-server"
+    "test": "nodemon src/main.js"
   },
   "repository": {
     "type": "git",
-    "url": ""
+    "url": "git+https://github.com/luctst/crispy-octo-adventure.git"
   },
-  "author": "Tostée Lucas <lucas.tostee@gmail.com> (https://www.lucas-tostee.com)",
-  "license": "",
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "bugs": {
-    "url": ""
+    "url": "https://github.com/luctst/crispy-octo-adventure/issues"
   },
-  "homepage": "",
-  "dependencies": {},
+  "homepage": "https://github.com/luctst/crispy-octo-adventure#readme",
   "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.2.3",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.4",
-    "css-loader": "^2.1.0",
-    "file-loader": "^3.0.1",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^4.11.0",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "sass-loader": "^7.1.0",
-    "webpack": "^4.28.4",
-    "webpack-cli": "^3.2.1",
-    "webpack-dashboard": "^2.0.0",
-    "webpack-dev-server": "^3.1.14"
+    "autoprefixer": "^9.1.3",
+    "browser-sync": "^2.24.7",
+    "del": "^3.0.0",
+    "gulp": "^3.9.1",
+    "gulp-clean-css": "^3.10.0",
+    "gulp-htmlmin": "^4.0.0",
+    "gulp-imagemin": "^4.1.0",
+    "gulp-postcss": "^8.0.0",
+    "gulp-sass": "^4.0.1",
+    "hint": "^3.4.3",
+    "postcss-discard-duplicates": "^4.0.0",
+    "run-sequence": "^2.2.1"
   }
 }
