+++ src/index.html
@@ -6,9 +6,33 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
     <link rel="stylesheet" href="main.css">
-    <title>Web-timeline</title>
+    <title>Evolution of technologies</title>
 </head>
 <body>
+    <header>
+        <h1 class="is__main__title">The evolution of technologies</h1>
+    </header>
+    <nav class="filter" id="js-stickyMenu">
+        <div class="filter--date">
+            <h3 class="is__subTitle is__date">Date</h3>
+        </div>
+        <div class="filter--cat">
+            <select class="is__select is__subTitle">
+                <option value="">Categories</option>
+                <option value="Network">Network</option>
+                <option value="Launch">Launch</option>
+                <option value="Science">Science</option>
+                <option value="Security">Security</option>
+                <option value="Programming">Programming</option>
+                <option value="Ai">AI</option>
+                <option value="Social Media">Social Media</option>
+                <option value="design">Design</option>
+            </select>
+        </div>
+        <div class="filter--title">
+            <h3 class="is__subTitle">Title</h3>
+        </div>
+    </nav>
     <main id="app"></main>
     <script src="main.js"></script>
 </body>
