+++ src/sass/general.scss
@@ -1,4 +1,5 @@
 body{
+    font-family: "Roboto", Arial, Helvetica, sans-serif;
     max-width: 100%;
     margin: 0;
     padding: 0;
