+++ src/sass/helper/heritage.scss
@@ -4,6 +4,17 @@
 @import "variable.scss";
 
 /**
+* Button
+*/
+%btn {
+    border-radius: 50px 50px 50px 50px;
+    background: $primary-color;
+    width: 15%;
+    height: auto;
+    padding: 8px 55px;
+}
+
+/**
 * Grid layout
 */
 %grid--layout {
@@ -13,14 +24,6 @@
     justify-items: start;
     align-items: center;
     padding: 1em;
-}
-
-%btn {
-    width: 25%;
-    height: auto;
-    cursor: pointer;
-    -webkit-transition-duration: 0.4s; /* Safari */
-    transition-duration: 0.4s;
 }
 
 /**
