+++ src/sass/helper/variable.scss
@@ -3,6 +3,6 @@ $primary-color:#4D9DC2;
 $secondary-color: #4083A2;
 $background_important-color:#FFEFDE;
 $txt_primary-color:white;
-$txt_secondary-color: #4083A2;
+$txt_secondary-color: #333;
 $important-color: #FF7B0A;
 $table_bg-primary-color: #F2F7FA;
