--- src/view/components/Header.js
@@ -1,20 +1,0 @@
-/**
- * Import
- */
-import React from 'react';
-
-/**
- * Déclaration
- */
-const Header = () => {
-    return (
-        <header>
-            <h1 classNames="is__main__title">The evolution of technologies</h1>
-        </header>
-    );
-}
-
-/**
- * Export
- */
-export default Header;
